slide1 for home page created
