+++ src/index.js
@@ -1,7 +1,8 @@
-import { createHeader } from "./modules/header";
+import { createHome } from "./modules/home";
 require("./css/style.css");
 
 document.addEventListener(
   "DOMContentLoaded",
-  document.getElementById("content").appendChild(createHeader.header)
+  document.getElementById("content").appendChild(createHome.header),
+  document.getElementById("content").appendChild(createHome.slide)
 );
